% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateSpatialEstimates.R
\name{generate.spatial.estimates}
\alias{generate.spatial.estimates}
\title{Function to generate estimates of known group sizes}
\usage{
generate.spatial.estimates(
  params,
  population.names,
  unknown.indices,
  pop.col.name,
  group.info
)
}
\arguments{
\item{params}{list of posterior samples of rho and L}

\item{population.names}{vector of population names}

\item{unknown.indices}{vector of indices of unknown groups}

\item{pop.col.name}{name of column in group.info containing population sizes}

\item{group.info}{data frame of group information}
}
\value{
data frame of estimates of known group sizes
}
\description{
Function to generate estimates of known group sizes
}
