% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateSpatialEstimates.R
\name{generateSpatialEstimates}
\alias{generateSpatialEstimates}
\title{Function to generate estimates of known municipality sizes}
\usage{
generateSpatialEstimates(
  params,
  population.names,
  unknown.indices,
  pop.col.name,
  municipality.info
)
}
\arguments{
\item{params}{list of posterior samples of rho and L}

\item{population.names}{vector of population names same as column names of municipality.info}

\item{unknown.indices}{vector of indices of unknown municipalitys}

\item{pop.col.name}{name of column in municipality.info containing population sizes}

\item{municipality.info}{data frame of municipality information in the sample order as the ard passed to run.sampling}
}
\value{
data frame of estimates of group sizes for each municipality
}
\description{
Function to generate estimates of known municipality sizes
}
