% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posteriorRhoScaling.R
\name{posteriorRhoScaling}
\alias{posteriorRhoScaling}
\title{Function to scale posterior samples of rho by prior known group sizes}
\usage{
posteriorRhoScaling(rho, raw_known_sizes, population_size)
}
\arguments{
\item{rho}{negative of matrix of posterior samples of rho_j}

\item{raw_known_sizes}{vector of prior known group sizes for municipality}

\item{population_size}{total population size for municipality}
}
\value{
matrix of scaled posterior samples of rho_j
}
\description{
Function to scale posterior samples of rho by prior known group sizes
}
