% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runSampling.R
\name{run.sampling}
\alias{run.sampling}
\title{Function to sample from proposed distribution}
\usage{
run.sampling(
  ard,
  pop.sub.regions,
  grouping,
  num.iterations,
  num.chains,
  pars = c("rho_j", "L"),
  ...
)
}
\arguments{
\item{ard}{matrix of respondent data}

\item{pop.sub.regions}{vector of population sizes for subregions}

\item{grouping}{vector of group for each respondent}

\item{num.iterations}{number of iterations to run}

\item{num.chains}{number of chains to run}

\item{pars}{parameters to save. only extract necessary rho_j and L by default}

\item{...}{additional arguments to pass to stan like seed, cores, messages, etc.}
}
\value{
posterior samples of rho and L
}
\description{
Function to sample from proposed distribution
}
