% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runSampling.R
\name{runSampling}
\alias{runSampling}
\title{Function to sample from proposed distribution}
\usage{
runSampling(ard, grouping, num.iterations, num.chains, pars = c("rho_j"), ...)
}
\arguments{
\item{ard}{matrix of respondent data}

\item{grouping}{vector of group for each respondent}

\item{num.iterations}{number of iterations to run}

\item{num.chains}{number of chains to run}

\item{pars}{parameters to save. only extract necessary rho_j by default}

\item{...}{additional arguments to pass to stan like seed, cores, messages, etc.}
}
\value{
posterior samples of rho
}
\description{
Function to sample from proposed distribution
}
