% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subnationalPrev-package.R
\docType{package}
\name{subnationalPrev-package}
\alias{subnationalPrev-package}
\alias{subnationalPrev}
\title{The 'subnationalPrev' package.}
\description{
A DESCRIPTION OF THE PACKAGE
}
\references{
Stan Development Team (NA). RStan: the R interface to Stan. R package version 2.32.3. https://mc-stan.org
}
