% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateSpatialEstimates.R
\name{compute.correlation.samples}
\alias{compute.correlation.samples}
\title{Function to calculate the average correlation between groups}
\usage{
compute.correlation.samples(params)
}
\arguments{
\item{params}{posterior samples}
}
\value{
a matrix of average correlations between groups extracted from L
}
\description{
Function to calculate the average correlation between groups
}
