% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlatedScaling.R
\name{correlatedScaling}
\alias{correlatedScaling}
\title{Function to scale posterior samples of rho by prior known group sizes}
\usage{
correlatedScaling(rho, correlation, raw_known_sizes, population_size)
}
\arguments{
\item{rho}{matrix from posterior samples of rho_j}

\item{correlation}{matrix from posterior samples of L}

\item{raw_known_sizes}{vector of prior known group sizes for subregion}

\item{population_size}{total population size for subregion}
}
\value{
matrix of scaled posterior samples of rho_j
}
\description{
Function to scale posterior samples of rho by prior known group sizes
}
