% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runSampling.R
\name{run.sampling}
\alias{run.sampling}
\title{Function to sample from proposed distribution}
\usage{
run.sampling(
  ard,
  pop.sub.regions,
  grouping,
  num.iterations,
  num.cores,
  num.chains,
  seed
)
}
\arguments{
\item{ard}{matrix of respondent data}

\item{pop.sub.regions}{vector of population sizes for subregions}

\item{grouping}{vector of group for each respondent}

\item{num.iterations}{number of iterations to run}

\item{num.cores}{number of cores to use}

\item{num.chains}{number of chains to run}

\item{seed}{random seed}
}
\value{
posterior samples of rho and L
}
\description{
Function to sample from proposed distribution
}
